More refactoring of GraphQL scalars
