+++ datajunction-server/datajunction_server/api/graphql/scalars/metricmetadata.py
@@ -1,0 +1,32 @@
+"""Metric metadata scalars"""
+
+from typing import Optional
+
+import strawberry
+
+from datajunction_server.models.node import MetricDirection as MetricDirection_
+
+MetricDirection = strawberry.enum(MetricDirection_)
+
+
+@strawberry.type
+class Unit:  # pylint: disable=too-few-public-methods
+    """
+    Metric unit
+    """
+
+    name: str
+    label: Optional[str]
+    category: Optional[str]
+    abbreviation: Optional[str]
+    description: Optional[str]
+
+
+@strawberry.type
+class MetricMetadata:  # pylint: disable=too-few-public-methods
+    """
+    Metric metadata output
+    """
+
+    direction: Optional[MetricDirection]  # type: ignore
+    unit: Optional[Unit]
